Make Sysstat Service Skippable
When run under docker containers (as with Molecule), the default CentOS
docker container doesn't provide any service daemons (as it should!). So
when trying to manage services in the container, the tasks fail.

Adding this option allows users to skip enabling the sysstat service so
the role can be applied inside of a Docker container.

Change-Id: I2cc4bb47cd50e716e30fcf19b2bdeac80b25c9db
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>