Chore: Update pre-commit dependencies to latest
ERROR! the role 'sysstat-install' was not found in <>
The error appears to be in '': line 5, column 7, but may
be elsewhere in the file depending on the exact syntax
problem.

Update role name to match the role path on GHA and
Jenkins jobs.

Add GHA verify workflow for verify jobs.
Update tox and pre-commit dependencies.

Ref: https://github.com/ansible/molecule/issues/1042
Change-Id: If512a38b9cc2ed94a8931a4d7e0a1409f95666af
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>